fix: better progress display for second-order samplers
Second-order samplers call the model twice per step, except
for the last iteration. Since the progress is updated only
for each second call, the last step is never shown.

The timing information is also misleading, since it's
displaying the number of full steps, but measuring only
the last half. Comparing to a first-order sampler, the
progress shows the same timing for each iteration, and
the same number of steps, but takes almost twice as long.

So, change the display to show average time per step,
which should give a better idea of the expected time
until completion, and update the progress display after
all model calls.